Removed puts and renamed gets to prompt.
The puts method doesn't really solve a problem that needs
async.js so I have removed it. If you liked the minimalistic
log it used, you'll be happy to know that I'm currently
making a whole new library based on it.

Gets usually means to just get input from the user. Prompt
is a better name as it implies getting input from the user
after asking them a question or stating something.